# docs/swagger.yaml
definitions:
  notes.Note:
    properties:
      author:
        type: string
      content:
        type: string
      id:
        type: integer
    type: object
  users.User:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: API for Studysyncr practice project
  title: Studysyncr API
  version: "1.0"
paths:
  /authorise:
    post:
      consumes:
      - application/json
      description: Authorise a user
      parameters:
      - description: user JSON
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/users.User'
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Authorisation
      tags:
      - authorisation
  /notes:
    post:
      consumes:
      - application/json
      description: Upload a new note
      parameters:
      - description: Note JSON
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/notes.Note'
      - description: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwidXNlcm5hbWUiOiJ1c2VyMSJ9.f3FtHCssZ2xJN74rmYeautqiMDrejzBbnspf222cfbo
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notes.Note'
      summary: Post a note
      tags:
      - notes
  /notes/:
    get:
      description: Returns JSON array of notes
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notes.Note'
            type: array
      summary: Get all notes of user
      tags:
      - notes
  /notes/{id}:
    delete:
      description: Deletes a note from db
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: note id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notes.Note'
      summary: Delete a note
      tags:
      - notes
    get:
      description: Returns JSON of note
      parameters:
      - description: note id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notes.Note'
      summary: Get a note
      tags:
      - notes
    patch:
      consumes:
      - application/json
      description: Updates a note in the db
      parameters:
      - description: note id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Note JSON
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/notes.Note'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notes.Note'
      summary: Patch a note
      tags:
      - notes
  /register:
    post:
      consumes:
      - application/json
      description: Register a user
      parameters:
      - description: user JSON
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Register user
      tags:
      - registration
swagger: "2.0"